refactor(extension): typescript support file
