+++ package.json
@@ -49,8 +49,11 @@
     "trailingComma": "none"
   },
   "devDependencies": {
+    "@types/node": "^15.12.4",
+    "@types/react": "^17.0.11",
     "autoprefixer": "^10.2.6",
     "postcss": "^8.3.0",
-    "tailwindcss": "^2.1.3"
+    "tailwindcss": "^2.1.3",
+    "typescript": "^4.3.4"
   }
 }
