+++ src/layouts/default.tsx
@@ -3,7 +3,6 @@ import { motion } from 'framer-motion'
 import { Menu } from '@headlessui/react'
 import ActiveLink from '~/utils/ActiveLink'
 import { BsThreeDots } from 'react-icons/bs'
-import { useState, useCallback } from 'react'
 import Scrollbar from 'react-smooth-scrollbar'
 import ThemeChanger from '~/utils/ThemeChanger'
 import { LogoIcon, SoundIcon } from '~/utils/Icons'
@@ -11,9 +10,6 @@ import { navigations, socialLinks } from '~/static/default'
 import MessengerCustomerChat from 'react-messenger-customer-chat'
 
 export default function Layout ({ children }) {
-  const [isAudio, setIsAudio] = useState() 
-  const toggleAudio = useCallback(() => setIsAudio(v => !v), [])
-
   return (
     // Main Layout
     <div className="font-sans antialiased flex flex-col md:flex-row justify-between w-full min-h-screen h-screen bg-white text-black dark:bg-dark-dim dark:text-white transition ease-in-out duration-700">
@@ -21,15 +17,19 @@ export default function Layout ({ children }) {
       {/* First Flex Column */}
       <div className="flex items-center flex-row md:flex-col justify-between h-auto md:h-full pb-3 border-b border-gray-200 dark:border-gray-600 md:border-0 px-4 py-4 md:py-12 md:px-10">
         <div className="flex items-center space-x-2">
-          <ActiveLink href="/" current="">
+          <ActiveLink 
+            href="/" 
+            default=""
+            current=""
+          >
             <a>
               <div className="flex-shrink-0">
                 <LogoIcon className="w-8 h-8 fill-current transform rotate-90" />
               </div>
             </a>
           </ActiveLink>
-          <button onClick={toggleAudio} className="focus:outline-none">
-            <SoundIcon isAudio={isAudio} className="w-6 h-6 text-gray-200 dark:text-gray-500" />
+          <button className="focus:outline-none">
+            <SoundIcon className="w-6 h-6 text-gray-200 dark:text-gray-500" />
           </button>
         </div>
         <div className="hidden md:block">
@@ -77,7 +77,11 @@ export default function Layout ({ children }) {
             <ul className="flex items-center justify-between space-x-0 md:space-x-8">
               {navigations.map(({ text, href }, i) => (
                 <li key={i} className="w-1/4 text-center">
-                  <ActiveLink href={href} current="font-semibold text-blue-twitter">
+                  <ActiveLink 
+                    href={href} 
+                    default=""
+                    current="font-semibold text-blue-twitter"
+                  >
                     <a className="text-xs tracking-widest hover:text-blue-twitter transition ease-in-out duration-200">
                       { text }
                     </a>
