+++ src/static/default.tsx
