+++ src/static/projects.tsx
